+++ fpl_common/src/types/player_id.rs
@@ -6,17 +6,23 @@ use std::ops::Deref;
     Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Deserialize, Serialize, sqlx::Type,
 )]
 #[sqlx(transparent)]
-pub struct PlayerId(pub i32);
+pub struct PlayerId(pub i16);
 
 impl PlayerId {
-    pub const fn new(id: i32) -> Self {
+    pub const fn new(id: i16) -> Self {
         Self(id)
     }
 
-    pub fn as_i32(&self) -> i32 {
+    pub fn as_i32(&self) -> i16 {
         self.0
     }
 }
+
+// impl Default for PlayerId {
+//     fn default() -> Self {
+//         Self(0)
+//     }
+// }
 
 impl Display for PlayerId {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
@@ -24,20 +30,20 @@ impl Display for PlayerId {
     }
 }
 
-impl From<i32> for PlayerId {
-    fn from(id: i32) -> Self {
+impl From<i16> for PlayerId {
+    fn from(id: i16) -> Self {
         Self(id)
     }
 }
 
-impl From<PlayerId> for i32 {
+impl From<PlayerId> for i16 {
     fn from(id: PlayerId) -> Self {
         id.0
     }
 }
 
 impl Deref for PlayerId {
-    type Target = i32;
+    type Target = i16;
 
     fn deref(&self) -> &Self::Target {
         &self.0
